+++ components/dial/stories/dial.stories.js
@@ -2,6 +2,7 @@ import { disableDefaultModes } from "@spectrum-css/preview/modes";
 import { isFocused } from "@spectrum-css/preview/types";
 import { version } from "../package.json";
 import { DialGroup } from "./dial.test.js";
+import { Template } from "./template.js";
 
 /**
  * A dial is an input control used for selecting a value within a range, similar to a slider. It's often used in audio and video mixing and editing applications, where horizontal space is limited.
@@ -28,7 +29,7 @@ export default {
 			},
 			control: "text",
 		},
-		isFocused,
+		isFocusVisible: isFocused,
 		isDragged: {
 			name: "Dragged",
 			type: { name: "boolean" },
@@ -51,7 +52,7 @@ export default {
 	args: {
 		rootClass: "spectrum-Dial",
 		size: "m",
-		isFocused: false,
+		isFocusVisible: false,
 		isDragged: false,
 		isDisabled: false,
 	},
@@ -62,6 +63,33 @@ export default {
 
 export const Default = DialGroup.bind();
 Default.args = {};
+
+export const Small = Template.bind();
+Small.args = {
+	size: "s",
+};
+Small.parameters = {
+	chromatic: { disableSnapshot: true },
+};
+Small.tags = ["!dev"];
+
+export const WithLabel = Template.bind();
+WithLabel.args = {
+	label: "Volume",
+};
+WithLabel.parameters = {
+	chromatic: { disableSnapshot: true },
+};
+WithLabel.tags = ["!dev"];
+
+export const Disabled = Template.bind();
+Disabled.args = {
+	isDisabled: true,
+};
+Disabled.parameters = {
+	chromatic: { disableSnapshot: true },
+};
+Disabled.tags = ["!dev"];
 
 // ********* VRT ONLY ********* //
 export const WithForcedColors = DialGroup.bind({});