+++ components/dialog/stories/template.js
@@ -15,9 +15,8 @@ import "../themes/spectrum.css";
 
 export const Template = ({
 	rootClass = "spectrum-Dialog",
-	isDismissable = true,
-	isFullScreen = false,
-	isFullScreenTakeover = false,
+	isDismissible = false,
+	hasDivider = true,
 	isOpen = true,
 	showModal = false,
 	hasFooter = false,
@@ -26,6 +25,10 @@ export const Template = ({
 	footer = [],
 	customClasses = [],
 	id = getRandomId("dialog"),
+	size = "medium",
+	layout,
+	heroImageUrl,
+	customStyles = {},
 } = {}, context = {}) => {
 	const { globals = {}, updateArgs } = context;
 
@@ -38,54 +41,68 @@ export const Template = ({
 		<div
 			class=${classMap({
 				[rootClass]: true,
-				[`${rootClass}--${scale}`]: true,
-				[`${rootClass}--dismissable`]: isDismissable && !isFullScreen && !isFullScreenTakeover,
-				[`${rootClass}--fullscreen`]: isFullScreen,
-				[`${rootClass}--fullscreenTakeover`]: isFullScreenTakeover,
+				[`${rootClass}--dismissable`]: isDismissible && ["fullscreen", "fullscreenTakeover"].every(l => layout !== l),
+				[`${rootClass}--${layout}`]: typeof layout !== "undefined",
+				[`${rootClass}--${size}`]: typeof size !== "undefined", 
+				[`${rootClass}--noDivider`]: !hasDivider,
 				...customClasses.reduce((a, c) => ({ ...a, [c]: true }), {}),
 			})}
 			id=${ifDefined(id)}
 			role="dialog"
 			tabindex="-1"
 			aria-modal="true"
+			style=${ifDefined(styleMap(customStyles))}
 		>
 			<div class="${rootClass}-grid">
-				${when(heading, () => [
-					html`<h1 class="${rootClass}-heading">${heading}</h1>`,
+				${when(typeof heroImageUrl !== "undefined", () =>
+					html`
+						<div 
+							class="spectrum-Dialog-hero"
+							style="background-image:url(${heroImageUrl})">
+						</div>
+					`
+				)}
+				${when(heading, () => html`
+					<h1 class="${rootClass}-heading">${heading}</h1>
+				`)}
+				${when(hasDivider, () =>
 					Divider({
 						horizontal: true,
 						customClasses: [`${rootClass}-divider`],
 					}, context),
-				])}
+				)}
 				<section class="${rootClass}-content">
 					${renderContent(content)}
 				</section>
-				${when(hasFooter || footer, () => {
+				${when(hasFooter, () => {
 					return html`
 						<footer class="${rootClass}-footer" style=${styleMap({
 							"justify-content": "flex-end",
 						})}>
 							${when(typeof footer !== "undefined" && Array.isArray(footer) && footer.length > 0,
 								() => renderContent(footer),
-								() => ButtonGroup({
-									items: [
-										{
-											label: "Cancel",
-											treatment: "outline",
-											variant: "secondary",
-										},
-										{
-											label: "Save",
-											treatment: "fill",
-											variant: "accent"
-										},
-									],
-								}, context)
+							)}
+							${when(!isDismissible, () => html `
+									${ButtonGroup({
+										items: [
+											{
+												label: "Cancel",
+												treatment: "outline",
+												variant: "secondary",
+											},
+											{
+												label: "Save",
+												treatment: "fill",
+												variant: "accent"
+											},
+										],
+									}, context)}
+								`
 							)}
 						</footer>
 					`;
 				})}
-				${when(isDismissable && !isFullScreen && !isFullScreenTakeover, () =>
+				${when(isDismissible && layout !== "fullscreen" && layout !== "fullscreenTakeover", () =>
 					CloseButton({
 						customClasses: [`${rootClass}-closeButton`],
 						onclick: toggleOpen,
@@ -100,6 +117,7 @@ export const Template = ({
 			${Modal({
 				isOpen,
 				content: [ () => Dialog],
+				variant: layout,
 			}, context)}
 		`;
 	}
@@ -107,3 +125,4 @@ export const Template = ({
 		return Dialog;
 	}
 };
+