+++ components/divider/stories/template.js
@@ -13,6 +13,7 @@ export const Template = ({
 	tag = "hr",
 	staticColor,
 	vertical = false,
+	minDimensionValues,
 	customClasses = [],
 	customStyles = {},
 } = {}) => {
@@ -27,9 +28,8 @@ export const Template = ({
 				...customClasses.reduce((a, c) => ({ ...a, [c]: true }), {}),
 			})}
 			style=${styleMap({
-				"block-size": vertical ? "auto" : undefined,
-				"inline-size": !vertical ? "100%" : undefined,
-				"align-self": vertical == true ? "stretch" : undefined,
+				"min-inline-size": minDimensionValues && !vertical ? "200px": undefined,
+				"min-block-size": minDimensionValues && vertical ? "20px": undefined,
 				...customStyles,
 			})}
 			role="separator"
@@ -46,9 +46,8 @@ export const Template = ({
 			...customClasses.reduce((a, c) => ({ ...a, [c]: true }), {}),
 		})}
 		style=${styleMap({
-			"min-height": vertical == true ? "20px" : undefined,
-			height: vertical == true ? "auto" : undefined,
-			"align-self": vertical == true ? "stretch" : undefined,
+			"min-inline-size": minDimensionValues && !vertical ? "200px": undefined,
+			"min-block-size": minDimensionValues && vertical ? "20px": undefined,
 			...customStyles,
 		})}
 		role="separator"