+++ components/dropindicator/stories/template.js
@@ -29,3 +29,20 @@ export const Template = ({
 		></div>
 	`;
 };
+
+export const DocsDropIndicatorGroup = (args, context) => html`
+	<div style=${styleMap({
+		"display": "flex",
+		"align-items": "center",
+		"gap": "16px",
+		"justify-content": "space-evenly",
+	})}>
+	${Template({
+			...args,
+		}, context)}
+		${Template({
+			...args,
+			direction: "horizontal",
+		}, context)}
+	</div>
+`;