+++ components/logicbutton/stories/template.js
@@ -28,3 +28,11 @@ export const Template = ({
 		</button>
 	`;
 };
+
+export const VariantGroup = (args, context) => html`
+	${Template({...args}, context)}
+	${Template({
+		...args,
+		variant: "or",
+	}, context)}
+`;