+++ components/progressbar/stories/progressbar.stories.js
@@ -57,6 +57,8 @@ export default {
 			control: { type: "range", min: 0, max: 100,},
 			if: { arg: "isIndeterminate", truthy: false },
 		},
+		trackFill: { table: { disable: true }},
+		progressBarFill: { table: { disable: true }},
 		staticColor: {
 			name: "Static color",
 			type: { name: "string" },